Create STOCK PRICE PREDICTION
Predicting the stock market has been the bane and goal of investors since its inception. Every day billions of dollars are traded on the stock exchange, and behind every dollar is an investor hoping to make a profit in one way or another. Here, your challenge is to build a predictive model that can predict stock prices.